feat: create tftpc binary
