+++ .github/workflows/unit.yml
@@ -17,5 +17,9 @@ jobs:
       - uses: actions/checkout@v3
       - name: Build
         run: cargo build --verbose
+      - name: Build
+        run: cargo build --features client --verbose
       - name: Run tests
         run: cargo test --verbose
+      - name: test client
+        run: cargo test --features client --verbose
