+++ README.md
@@ -34,28 +34,22 @@ tftpd -i 0.0.0.0 -p 1234 -d "/home/user/tftp" -r
 
 ## Usage (Client)
 
+Client code is protected by a feature flag names `client`.
 To install the client and server using Cargo:
 
 ```bash
 cargo install --features client tftpd
-tftpd client --help
-tftpd server --help
-```
-
-To run the server on the IP address `0.0.0.0`, read-only, on port `1234` in the `/home/user/tftp` directory:
-
-```bash
-tftpd server -i 0.0.0.0 -p 1234 -d "/home/user/tftp" -r
+tftpc --help
 ```
 
 To connect the client to a tftp server running on IP address `127.0.0.1`, read-only, on port `1234` and download a file named `example.file`
 ```bash
-tftpd client example.file -i 0.0.0.0 -p 1234 -d
+tftpc example.file -i 0.0.0.0 -p 1234 -d
 ```
 
 To connect the client to a tftp server running on IP address `127.0.0.1`, read-only, on port `1234` and upload a file named `example.file`
 ```bash
-tftpd client ./example.file -i 0.0.0.0 -p 1234 -u
+tftpc ./example.file -i 0.0.0.0 -p 1234 -u
 ```
 
 ## License
